--- src/components/AddressInfo/styles.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  addressContainer: {
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  addressText: {
-    fontSize: 15,
-    lineHeight: 18,
-    color: 'black',
-    textAlign: 'center',
-  },
-});
